javadoc, conversion of some public functions to private functions and change a package name from tsp to algorithm
