+++ src/main/java/fr/insa/lyon/pld/agile/algorithm/TSPSolverFactory.java
@@ -1,4 +1,4 @@
-package fr.insa.lyon.pld.agile.tsp;
+package fr.insa.lyon.pld.agile.algorithm;
 
 public abstract class TSPSolverFactory {
